pyproject.toml
[tool.poetry]
name = "clubmanager"
version = "0.0.1"
description = "Club Management software"
authors = ["Bernard Siebens <bernard@siebens.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.7"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-auditlog = "^3.0.0"
django-compressor = "^4.5.1"
django-initials-avatar = "^0.0.7"
django-extensions = "^3.2.3"
django-filter = "^24.3"
django-debug-toolbar = "^4.4.6"
rules = "^3.4"
django-markdownx = "^4.0.7"
django-cleanup = "^8.1.0"
python-dateutil = "^2.9.0"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.21.4"
django-cors-headers = "^4.4.0"
beautifulsoup4 = "^4.12.3"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
flower = "^2.0.1"
django-model-utils = "^4.5.1"
django-two-factor-auth = {extras = ["webauthn"], version = "^1.17.0"}
django-environ = "^0.11.2"
pillow = "^10.4.0"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
flake8-django = "^1.4"
pytest = "^8.3.2"
coverage = "^7.6.0"
pytest-django = "^4.8.0"
teamcity-messages = "^1.32"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
